Implement Chris' localization suggestion, and make file entries optional
It is now only necessary to specify an image filename in metadata.yml
if you really want to have a specific caption or ordering for an image.
Updated manual to make this clear.
Update galleries.py to make it an error to provide a YAML entry which
does not have a name: field.