Replace PyYAML with ruamel.yaml.
PyYAML is effectively unmaintained. Many sources are considering PyYAML
insecure, and even though this claim is debatable, they still want the
unsafe API to be default. Ruamel’s fork is actively maintained and has a
saner approach. In addition, if the YAML-config project eventually
becomes a thing, we would need that library anyway.

Signed-off-by: Chris Warrick <kwpolska@gmail.com>