gallery was losing ordering
sorry for the quick and dirty PR.
It seems that the sorting was actually getting lost in older python versions. IIRC, python 3.6 introduced `dict`s that retain the insertion order, but it shouldn't really be relied on. I'm on python 3.5 though, and the dict implementation was jumbeling up the sort that happens a few lines before as it seems.