Use new importlib features to import config
Issue #3314 revealed that the new config import method using `imp`,
introduced in #3244, was broken if the config file imported other
stuff. An attempt to fix this (commit 37be3063c) led to unusual,
although mostly consistent, test failures, caused by sites building
partially. Moreover, the `imp` module was deprecated in Python 3.4.
New versions of the docs suggest that the replacement is
`importlib.util.module_from_spec`, which was introduced in 3.5. The
new implementation of config file loading does everything correctly:
the broken integration tests pass, the site on which #3314 manifested
builds, and `test_config.py` (which tests the new features, and would
fail in v8.0.2) passes as well.