Add support for the utterances comment system (https://utteranc.es)
Fixes #3453

There is one issue with this, which is that the configuration options for
Utterances are stored in GLOBAL_CONTEXT['utterances_config'], similar to isso,
but one configuration option is required (either issue-term or issue-number).
How can we make this error in the build if it isn't provided. Otherwise, the
only indication that it isn't working is an error in the console log.