Fix exception from gallery RSS with WebP images
WebP images in a gallery cause `nikola build` to fail with an unhandled
AttributeError exception, as described in #3671.

This happens because the MIME type of WebP images is not known, and the
resulting None value causes the RSS generation code to choke.

This PR fixes that, by augmenting Python's list of known mimetypes with
a hardcoded WebP one, before we ask for the mimetype of each image in a
gallery.

This addresses the above issue's checklist item "*Inject a hardcoded
'image/web' mimetype at runtime, before .guess_type() gets called.*",
but does not complete the issue.