Generate the webp again with correct exif data
The webp was previously generated by exporting from GIMP. This didn't
copy many of the EXIF metadata fields from the replaced jpg. This caused
a problem in the next PR, as a new test for RSS content is sensitive to
items like the image creation time, which is read from the EXIF
DateTimeDigitized.

Maybe GIMP can be cajoled into preserving this EXIF data, but instead,
this time I did the conversion using command line tool 'cwebp' (from
apt package webp):

    cwebp tesla_conducts_lg.jpg -metadata all -o tesla_conducts_lg.webp