Fix expected RSS content
The publish date on the webp image is now different,
since we rebased onto a change in the preceding branch,
which correctly populates this EXIF field on the new
webp file.

This affects the order if the items in the RSS feed,
and the title of the RSS feed. (possibly that last part
is a bug? But it is not today's problem.)