Gallery RSS test get XML description from subtitle
The new gallery RSS test can get access to the 'description' field
in the feed XML. It is a field labelled 'subtitle' in the parsed
feed data.

Delete the TODO related to this item.

Delete some debug printing.

Add a TODO wondering whether the gallery description should be blank,
or should be populated with the content from the gallery index.txt.