Use forward slashes in RSS item id on Windows
Windows RSS item ids are now of same form as on other OSes, ie:

    'galleries/demo/tesla_tower1_lg.jpg'

Formerly these used backslashes on Windows, which caused a bug where
excluded images (listed in exclude.meta), were not excluded because
their id didn't match.

This means I've removed a big ugly TODO from the test, which no
longer expects these excluded items to appear in the RSS feed on
Windows.

This is a re-application of the 'tentative fix' that
was applied and reverted earlier in this PR.